+++ NuRadioReco/modules/channelBandPassFilter.py
@@ -1,37 +1,8 @@
 import numpy as np
 from scipy import signal
-import scipy.signal
 import logging
-
 from NuRadioReco.modules.base.module import register_run
-from NuRadioReco.utilities import units
-from NuRadioReco.detector import filterresponse
-
-
-def get_filter_response(frequencies, passband, filter_type, order):
-    if (filter_type == 'rectangular'):
-        f = np.ones_like(frequencies)
-        f[np.where(frequencies < passband[0])] = 0.
-        f[np.where(frequencies > passband[1])] = 0.
-        return f
-    elif (filter_type == 'butter'):
-        f = np.zeros_like(frequencies, dtype=np.complex)
-        mask = frequencies > 0
-        b, a = scipy.signal.butter(order, passband, 'bandpass', analog=True)
-        w, h = scipy.signal.freqs(b, a, frequencies[mask])
-        f[mask] = h
-        return f
-    elif (filter_type == 'butterabs'):
-        f = np.zeros_like(frequencies, dtype=np.complex)
-        mask = frequencies > 0
-        b, a = scipy.signal.butter(order, passband, 'bandpass', analog=True)
-        w, h = scipy.signal.freqs(b, a, frequencies[mask])
-        f[mask] = h
-        return np.abs(f)
-    elif (filter_type.find('FIR') >= 0):
-        raise NotImplementedError("FIR filter not yet implemented")
-    else:
-        return filterresponse.get_filter_response(frequencies, filter_type)
+from NuRadioReco.utilities import units, filter
 
 
 class channelBandPassFilter:
@@ -157,7 +128,7 @@ class channelBandPassFilter:
             the complex filter amplitudes
         """
         tmp_passband, tmp_order, tmp_filter_type = self.get_filter_arguments(channel_id, passband, filter_type, order)
-        return get_filter_response(frequencies, tmp_passband, tmp_filter_type, tmp_order)
+        return filter.get_filter_response(frequencies, tmp_passband, tmp_filter_type, tmp_order)
 
     def _apply_filter(self, channel, passband, filter_type, order, is_efield=False):
